+++ pagerduty/resource_pagerduty_automation_actions_action.go
@@ -100,10 +100,13 @@ func resourcePagerDutyAutomationActionsAction() *schema.Resource {
 				Optional: true,
 			},
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 <<<<<<< Updated upstream
 =======
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+>>>>>>> 75344b3e (address rebase issues)
 			"only_invocable_on_unresolved_incidents": {
 				Type:     schema.TypeBool,
 				Computed: true,
@@ -121,8 +124,11 @@ func resourcePagerDutyAutomationActionsAction() *schema.Resource {
 				Computed: true,
 				Optional: true,
 			},
+<<<<<<< HEAD
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+>>>>>>> 75344b3e (address rebase issues)
 		},
 	}
 }
@@ -178,10 +184,13 @@ func buildAutomationActionsActionStruct(d *schema.ResourceData) (*pagerduty.Auto
 	}
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 <<<<<<< Updated upstream
 =======
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+>>>>>>> 75344b3e (address rebase issues)
 	if attr, ok := d.GetOk("only_invocable_on_unresolved_incidents"); ok {
 		val := attr.(bool)
 		automationActionsAction.OnlyInvocableOnUnresolvedIncidents = &val
@@ -208,8 +217,11 @@ func buildAutomationActionsActionStruct(d *schema.ResourceData) (*pagerduty.Auto
 	attr, _ := d.Get("allow_invocation_from_event_orchestration").(bool)
 	automationActionsAction.AllowInvocationFromEventOrchestration = &attr
 
+<<<<<<< HEAD
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+>>>>>>> 75344b3e (address rebase issues)
 	return &automationActionsAction, nil
 }