+++ pagerduty/resource_pagerduty_automation_actions_action_test.go
@@ -58,6 +58,7 @@ func TestAccPagerDutyAutomationActionsActionTypeProcessAutomation_Basic(t *testi
 					resource.TestCheckResourceAttrSet("pagerduty_automation_actions_action.foo", "runner_id"),
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "runner_type", "runbook"),
 <<<<<<< HEAD
+<<<<<<< HEAD
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "true"),
 =======
 <<<<<<< Updated upstream
@@ -67,6 +68,11 @@ func TestAccPagerDutyAutomationActionsActionTypeProcessAutomation_Basic(t *testi
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "true"),
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "true"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_from_event_orchestration", "true"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "true"),
+>>>>>>> 75344b3e (address rebase issues)
 				),
 			},
 			{
@@ -95,6 +101,7 @@ func TestAccPagerDutyAutomationActionsActionTypeProcessAutomation_Basic(t *testi
 					resource.TestCheckResourceAttrSet("pagerduty_automation_actions_action.foo", "runner_id"),
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "runner_type", "runbook"),
 <<<<<<< HEAD
+<<<<<<< HEAD
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
 =======
 <<<<<<< Updated upstream
@@ -104,6 +111,11 @@ func TestAccPagerDutyAutomationActionsActionTypeProcessAutomation_Basic(t *testi
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_from_event_orchestration", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
+>>>>>>> 75344b3e (address rebase issues)
 				),
 			},
 		},
@@ -148,6 +160,7 @@ func TestAccPagerDutyAutomationActionsActionTypeScript_Basic(t *testing.T) {
 					resource.TestCheckNoResourceAttr("pagerduty_automation_actions_action.foo", "runner_type"),
 					resource.TestCheckNoResourceAttr("pagerduty_automation_actions_action.foo", "runner_id"),
 <<<<<<< HEAD
+<<<<<<< HEAD
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
 =======
 <<<<<<< Updated upstream
@@ -157,6 +170,11 @@ func TestAccPagerDutyAutomationActionsActionTypeScript_Basic(t *testing.T) {
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_from_event_orchestration", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
+>>>>>>> 75344b3e (address rebase issues)
 				),
 			},
 			{
@@ -183,6 +201,7 @@ func TestAccPagerDutyAutomationActionsActionTypeScript_Basic(t *testing.T) {
 					resource.TestCheckNoResourceAttr("pagerduty_automation_actions_action.foo", "runner_type"),
 					resource.TestCheckNoResourceAttr("pagerduty_automation_actions_action.foo", "runner_id"),
 <<<<<<< HEAD
+<<<<<<< HEAD
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
 =======
 <<<<<<< Updated upstream
@@ -192,6 +211,11 @@ func TestAccPagerDutyAutomationActionsActionTypeScript_Basic(t *testing.T) {
 					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "only_invocable_on_unresolved_incidents", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_from_event_orchestration", "false"),
+					resource.TestCheckResourceAttr("pagerduty_automation_actions_action.foo", "allow_invocation_manually", "false"),
+>>>>>>> 75344b3e (address rebase issues)
 				),
 			},
 		},
@@ -282,18 +306,24 @@ resource "pagerduty_automation_actions_action" "foo" {
 	action_data_reference {
 		process_automation_job_id = "updated_pa_job_id_123"
 <<<<<<< HEAD
+<<<<<<< HEAD
 	}
 	only_invocable_on_unresolved_incidents = "false"
 =======
 <<<<<<< Updated upstream
 	  }
 =======
+=======
+>>>>>>> 75344b3e (address rebase issues)
 	}
 	only_invocable_on_unresolved_incidents = "false"
 	allow_invocation_manually = "false"
 	allow_invocation_from_event_orchestration = "false"
+<<<<<<< HEAD
 >>>>>>> Stashed changes
 >>>>>>> f352c045 (AA-1910 add allow_invocation_manually and allow_invocation_from_event_orchestration attributes to automation actions)
+=======
+>>>>>>> 75344b3e (address rebase issues)
 }
 `, previousActionName, actionName, actionDescription, actionClassification)
 }